unifying clusterId to be uint16

+++ waku/node/config.nim
@@ -12,7 +12,7 @@ import ../waku_enr
 
 type NetConfig* = object
   hostAddress*: MultiAddress
-  clusterId*: uint32
+  clusterId*: uint16
   wsHostAddress*: Option[MultiAddress]
   hostExtAddress*: Option[MultiAddress]
   wsExtAddress*: Option[MultiAddress]
@@ -78,7 +78,7 @@ proc init*(
     wssEnabled: bool = false,
     dns4DomainName = none(string),
     discv5UdpPort = none(Port),
-    clusterId: uint32 = 0,
+    clusterId: uint16 = 0,
     wakuFlags = none(CapabilitiesBitfield),
 ): NetConfigResult =
   ## Initialize and validate waku node network configuration
